refactor(orchestrator): extract LLM-call layer from executor_agent_loop
executor_agent_loop.go was 2252 LOC mixing the run loop, tool
dispatch, MCP routing, the in-loop HTTP fetch tool, telemetry,
and a small LLM-interaction layer. Pull the LLM-call helpers into
a sibling so the next reader can see "LLM stuff" and "tool stuff"
as distinct files.

  executor_agent_loop.go:      2252 → 2110 LOC
  executor_agent_loop_chat.go: new, 156 LOC

What moves to executor_agent_loop_chat.go

  - chatTurn: the streaming-or-not LLM call wrapper, including the
    persistPartial inner closure that incrementally updates the
    conversation artifact during streaming.
  - runWithoutLLM: the no-LLM short-circuit path for spec-time
    error cases (workspace missing, etc.).
  - isModelLacksToolsError: the upstream-tool-incapable error
    classifier consumed by Execute's main loop and used to flip
    to a failed-from-error result.
  - failedFromError: builds the final ExecutionResult for the
    failed run.

What stays in executor_agent_loop.go

  Construction (NewAgentLoopExecutor, SetMetrics, SetMCPHostFactory),
  the main Execute loop, tool dispatch (dispatchToolCall,
  dispatchMCPToolCall, runSubExecutor, gatedToolsInTurn, isGated,
  mcpServerPolicy), the in-loop HTTP fetch tool (httpRequestTool —
  next sibling target), the MCP telemetry recorder, the assistant-
  event emitters (emitAgentTurnStarted, emitAssistantMessageEvents,
  emitAssistantFinalAnswer), and step/artifact builders.

The receiver type stays *AgentLoopExecutor everywhere; no exported
symbol moves; tests untouched. Imports per file are minimal:
executor_agent_loop_chat.go takes context/strings/time +
internal/telemetry + pkg/types.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
